Settings Button funtional
Settings menu can be shut off by touching with another object. Does not yet toggle. Works using a collider in front of the button with an OnTriggerEnter() method called. The hand will not trigger an event, because the VR hands' collider components have "Is Trigger" enabled. Must find a way to fire an event with the hand without a trigger-button. To test, grab the source or a meter to touch the button.